refactor: make folder sorting explicit
